Adding service setting and group size flags
I added flags for entries that are tier II and individual setting, and
another flag for entries that are tier III and group. I created another
variable in ServiceD that is the unambiguous service setting (the
setting if tier and recorded.as match) and I created a flag for entries
where this unambiguous setting doesn’t match the group size that I
calculated using note, date, service provider combinations. There were
about 4% of observations where those didn’t match, so I’m thinking I
might need to change the way I calculate group size. Currently I am
calculating group size without looking at whether tier or recorded.as
indicate a group setting. I’m thinking we might want to change that so
that if there is an unambiguous individual setting from tier and
recorded.as there is automatically a group size of 1, and then group
size is calculated for all other services. The downside to that is it
is assuming that there are individual services that are being recorded
with the exact same note on the exact same day, which I know Sheri
doesn’t want… but at least the way we calculate group size and time
spent would be consistent with what the graduation coaches are telling
us. Let me know if that makes sense and you agree.