Adding service variables, changing group size calculation
I moved the group size calculation so that it is done after I create
the unambiguous “setting” variable, and automatically made all group
sizes 1 when setting is “Individual”. I added a sheet to our data
checks for tier/recorded.as mismatch flags.

I added 7 variables to studentlist that are aggregates of specific
kinds of services. tier1 is the sum of Tier I service hours, checkin is
the sum of student check in service hours, parent1on1 is the sum of
hours for 1-on-1 communication with parents, and anyfamily is the sum
of hours for any family contacts (1-on-1 parent contacts + family
focused events). num_check, num_parent1on1, and num_anyfamily are the
corresponding number of services delivered in these categories.